Modularize other parts of data processing
